requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
pydantic>=2.11,<3.0
pydantic-settings>=2.8,<3.0
sqlalchemy>=2.0,<3.0
alembic>=1.15,<2.0
PyJWT>=2.8,<2.9
psycopg[binary]>=3.2,<4.0
python-multipart>=0.0.9,<1.0
pytest>=8.3,<9.0
httpx>=0.28,<1.0
oss2>=2.18,<3.0
zhipuai>=2.0,<3.0
qdrant-client>=1.7,<2.0
anthropic>=0.40,<1.0
pypdf>=5.4,<7.0
redis>=5.0,<6.0
